[FE][Feat] #106 : context 이용해서 footer 버튼 활성화 기능 구현
